fixed strict types

+++ ZaehlerUeberlauf/module.php
@@ -37,7 +37,7 @@ declare(strict_types=1);
             }
         }
 
-        private function Update(int $OldValue, int $Value)
+        private function Update(float $OldValue, float $Value)
         {
             if (($Value - $OldValue) < 0) {
                 $diff = $this->ReadPropertyInteger('MaximumValue') + 1 - $OldValue + $Value;
